--- src/Characters/Mage.java
@@ -1,17 +1,0 @@
-package Characters;
-
-import Attributes.PrimaryAttribute;
-
-public class Mage extends Character{
-
-
-    public Mage(String name) {
-        super(name,
-                new PrimaryAttribute(1,1,8));
-    }
-
-    @Override
-    void levelUp() {
-
-    }
-}
