--- src/Characters/Ranger.java
@@ -1,16 +1,0 @@
-package Characters;
-
-import Attributes.PrimaryAttribute;
-
-public class Ranger extends Character{
-
-    public Ranger(String name) {
-        super(name,
-                new PrimaryAttribute(1,7,1));
-    }
-
-    @Override
-    void levelUp() {
-
-    }
-}
