+++ src/com/mod/assignment1rpgcharacters/Characters/Slot.java
@@ -1,4 +1,4 @@
-package Characters;
+package com.mod.assignment1rpgcharacters.Characters;
 
 public enum Slot {
         HEAD, BODY, ARMS, LEGS, WEAPON
